backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
python-multipart>=0.0.6          # File upload support

# ── LLM Council ───────────────────────────────────────────────────────────────
groq>=0.9.0                       # Groq Cloud API client
httpx>=0.27.0                     # Async HTTP (used by Groq SDK)

# ── ML / RAG ──────────────────────────────────────────────────────────────────
numpy>=1.26.0
scikit-learn>=1.4.0               # TF-IDF vectoriser + Logistic Regression
faiss-cpu>=1.7.0                  # FAISS vector similarity search

# ── Report Processing ────────────────────────────────────────────────────────
PyPDF2>=3.0.0                     # PDF text extraction
python-docx>=1.0.0                # DOCX text extraction

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
